Jest + testchain (#575)
* always run ganache & deploy contracts before jest
* set up web3 provider in test env
* remove unused code from config/networks
* get useVaultFactory test running
* support ganache & hardhat at once with different chain ids & ports
* always look up decimals from token contract or vault data
* add workaround for looking up some token addresses on mainnet fork
* allow running rebuild-hardhat when a testchain's already running

+++ scripts/rebuild-hardhat.sh
@@ -3,23 +3,31 @@ set -e
 
 FIND_DEV_PID="lsof -t -i:8545 -sTCP:LISTEN"
 
-git submodule update --init --recursive
-yarn hardhat:install --frozen-lockfile
-yarn hardhat:compile
+while [[ "$#" > 0 ]]; do case $1 in
+  --full) FULL=1;;
+esac; shift; done
 
-echo "Starting node for auto deployment..."
-yarn hardhat:dev >/dev/null &
+if [ "$FULL" ]; then
+  git submodule update --init --recursive
+  yarn hardhat:install --frozen-lockfile
+  yarn hardhat:compile
 
-while [ ! $(eval $FIND_DEV_PID) ]; do
-  sleep 1
-done
+  echo "Starting node for auto deployment..."
+  yarn hardhat:dev >/dev/null &
 
-yarn hardhat:deploy
+  while [ ! $(eval $FIND_DEV_PID) ]; do
+    sleep 1
+  done
+  yarn hardhat:deploy
+fi
+
 yarn hardhat:codegen
 yarn hardhat:build
 
-echo "Stopping node. The 'Command failed' messages that follow are normal."
-kill $(eval $FIND_DEV_PID)
+if [ "$FULL" ]; then
+  echo "Stopping node. The 'Command failed' messages that follow are normal."
+  kill $(eval $FIND_DEV_PID)
+fi
 
 cd hardhat
 yarn unlink >/dev/null 2>&1 || echo -n
